Added room creation
I messed up the contract with the server. Have to create a room before you can join it. So the gadget does that on connect now.

Removed a couple bits of test code that came along for the ride.

At this point I can join and leave rooms. Need to add the ability to test when I'm alone.